Refactor functions into services
